pyproject.toml
[project]
name = "ha-smartchain"
version = "0.8.0"
description = "SmartChain — multi-provider LLM conversation agent for Home Assistant"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"custom_components/smartchain/const.py" = ["E501"]

[dependency-groups]
dev = [
    "pytest-homeassistant-custom-component>=0.13",
    "ruff>=0.15",
    "langchain-core>=0.3,<1",
    "langchain-gigachat>=0.3.0",
    "langchain-openai>=0.3.0,<1",
    "langchain-community>=0.3.0,<0.4",
    "langchain-anthropic>=0.3.0,<1",
    "langchain-ollama>=0.3.0,<1",
    "gigachat>=0.1",
    "PyTurboJPEG>=2.0",
    "hassil>=3.5.0",
    "home-assistant-intents>=2026.3.3",
]
